refactor: removing release heading module file & type
- Too much bloat. Turned into a private module and a tuple as a result
of parsing instead.
