+++ src/parse/releases/mod.rs
@@ -1,11 +1,9 @@
 mod change_set;
 mod changes;
-mod heading;
 mod release;
 mod unreleased;
 
 pub use change_set::*;
 pub use changes::*;
-pub use heading::*;
 pub use release::*;
 pub use unreleased::*;
